add 3.1 version

+++ kandinsky3/model/nn.py
@@ -4,7 +4,7 @@ import torch
 from torch import nn, einsum
 from einops import rearrange, repeat
 
-from .utils import exist, set_default_layer
+from .utils import exist
 
 
 class Identity(nn.Module):
@@ -25,7 +25,7 @@ class SinusoidalPosEmb(nn.Module):
     def forward(self, x):
         half_dim = self.dim // 2
         emb = math.log(10000) / (half_dim - 1)
-        emb = torch.exp(torch.arange(half_dim, device=x.device) * -emb)
+        emb = torch.exp(torch.arange(half_dim, device=x.device, dtype=x.dtype) * -emb)
         emb = rearrange(x, 'i -> i 1') * rearrange(emb, 'j -> 1 j')
         return torch.cat((emb.sin(), emb.cos()), dim=-1)
